Refine ui
