+++ packages/comment-widget/src/index.ts
@@ -2,6 +2,7 @@ import { BaseCommentItem } from './base-comment-item';
 import { BaseCommentItemAction } from './base-comment-item-action';
 import { BaseForm } from './base-form';
 import { CommentItem } from './comment-item';
+import { CommentList } from './comment-list';
 import { CommentPagination } from './comment-pagination';
 import { CommentReplies } from './comment-replies';
 import { CommentWidget } from './comment-widget';
@@ -10,8 +11,8 @@ import { LoadingBlock } from './loading-block';
 import { ReplyForm } from './reply-form';
 import { ReplyItem } from './reply-item';
 import { UserAvatar } from './user-avatar';
-import './locale';
 
+import './locale';
 export {
   CommentWidget,
   BaseForm,
@@ -24,5 +25,6 @@ export {
   ReplyForm,
   ReplyItem,
   UserAvatar,
+  CommentList,
   LitToast,
 };
